api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
numpy
scikit-learn
google-generativeai